# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: LearningOS Platform API
  description: |
    Comprehensive API for India's AI-powered adaptive learning platform.
    Supports multi-tenant operations, DPI integrations (APAAR, DigiLocker, ABC),
    AI-driven personalization, and DPDP-compliant data governance.
  version: "1.0.0"
  contact:
    name: LearningOS Platform Team
    email: platform@learning-os.gov.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.platform.gov.in/v1
    description: Production
  - url: https://staging-api.platform.gov.in/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Local Development
security:
  - BearerAuth: []
...
paths:
...
  /dpi/abc/credits/deposit:
    post:
      tags: [DPI - ABC]
      summary: Deposit credits to ABC account
      operationId: depositCredits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditDepositRequest"
      responses:
        "200":
          description: Credits deposited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditDepositResponse"
  /dpi/abc/credits/transfer:
    post:
      tags: [DPI - ABC]
      summary: Transfer credits between institutions
      operationId: transferCredits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditTransferRequest"
      responses:
        "200":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditTransferResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreditDepositRequest:
      type: object
      required: [abcId, apaarId, credits, academicSession, consentId]
      properties:
        abcId:
          type: string
        apaarId:
          type: string
        credits:
          type: array
          items:
            type: object
            required: [courseId, credits, grade]
            properties:
              courseId:
                type: string
              credits:
                type: number
              grade:
                type: string
              term:
                type: string
              completedAt:
                type: string
                format: date-time
        academicSession:
          type: string
        consentId:
          type: string
    CreditDepositResponse:
      type: object
      properties:
        success:
          type: boolean
        referenceId:
          type: string
        totalCredits:
          type: number
        processedAt:
          type: string
          format: date-time
    CreditTransferRequest:
      type: object
      required: [abcId, apaarId, sourceInstitutionId, destinationInstitutionId, credits, consentId]
      properties:
        abcId:
          type: string
        apaarId:
          type: string
        sourceInstitutionId:
          type: string
        destinationInstitutionId:
          type: string
        credits:
          type: array
          items:
            type: object
            properties:
              courseId:
                type: string
              credits:
                type: number
              grade:
                type: string
              level:
                type: string
        consentId:
          type: string
    CreditTransferResponse:
      type: object
      properties:
        accepted:
          type: boolean
        transfer:
          type: object
          properties:
            transferId:
              type: string
            status:
              type: string
              enum: [pending, approved, rejected, completed]
        pendingActions:
          type: array
          items:
            type: string
...